Three more byte-decode corrections in expgfx, snowclaw and curves
gExpgfxStaticData is the plane-offset table the previous commit named,
so spell it f32[12] instead of 48 raw bytes; the values are -5.0 and
eleven 50.0. EXPGFX_STATIC_DATA still casts the base to the layout
overlay, which reaches past this object into the neighbouring carve.

gSnowClawMoveTable decodes as twelve big-endian words, not 48 bytes,
and seqPairTableLookup walks its argument as `int (*arr)[2]` comparing
[0] against a romDefNo and returning [1]. The pairs are
905/909, 906/910, 1235/1236, 365/364, 368/367, 1000/1002 - the
mount-partner romDefNo of each snowclaw variant. Typed s32.

PathControlInterface's two setter typedefs named their arguments
points/params/stride and pathDataA/pathDataB/pathParam. Their sole
implementations are curves_setLocalPointCollision and
curves_setSegmentCollision in engine/21, whose own parameters are
localPointPositions/localPointRadii/primaryHitType and
segmentLocalPoints/radii/types - and the fifth argument really is a
hit type (callers pass 1, 2, 4 and 8), not a stride.

Byte-identical: 1013 of 1013 source objects unchanged vs a pristine
build of the parent.

MEASURED AND DECLINED, so it is not re-probed: SpiritDoorLock's
`const u32 gSpiritDoorLockOrbitOffsetBase[4]` really does decode as
{0.0f, 0.0f, 7.0f, 0.0f}, but the SpiritDoorLockOrbitWords three-word
copy that seeds the f32[3] local is load-bearing - spelling the object
f32 and copying element-wise turns lwz/stw into lfs/stfs and takes the
unit from 100.0 to 98.986115. Retail copied words, so the object was
word-typed.

+++ include/main/expgfx_internal.h
@@ -472,7 +472,7 @@ STATIC_ASSERT(offsetof(ExpgfxSlot, renderX) == 0x90);
 #define EXPGFX_STATIC_DATA ((ExpgfxStaticDataLayout *)gExpgfxStaticData)
 #define EXPGFX_RUNTIME_DATA ((ExpgfxRuntimeDataLayout *)gExpgfxRuntimeData)
 
-extern u8 gExpgfxStaticData[];
+extern f32 gExpgfxStaticData[];
 extern u8 gExpgfxRuntimeData[];
 extern ExpgfxTableEntry gExpgfxTableEntries[];
 extern ObjAnimComponent* gExpgfxTrackedPoolSourceIds[];
